shells/happDev: Use hc-run-command to expose crate binaries
Consistently updating the expected hashs for the holochain Nix
derivation seems an unacceptable burden in everyday-development.
Still, we're looking for a way to expose the holochain binaries in PATH
somehow, without having to manually installing them.
The approach here is to put scripts in place that effectively wraps a
`cargo run` invocation to target a specific crate.

Example of how this can be used:

```
$ nix-shell --pure --argstr flavor happDev --run "holochain --version"
Using /home/steveej/src/holo/holochain_holochain as target prefix...
+ crate=holochain
+ shift
+ cargo run --target-dir=/home/steveej/src/holo/holochain_holochain --manifest-path=/home/steveej/src/holo/holochain_holochain/crates/holochain/Cargo.toml -- --version
    Finished dev [unoptimized + debuginfo] target(s) in 0.29s
     Running `debug/holochain --version`
holochain 0.0.1
```